Adds os-win master branch
We're going to test compute-hyperv with the master branch of os-win, to include
the latest changes in the library, including new features.

This is not done on nova or other projects. If there is a regression in os-win, it will
cause the CI to fail on those projects, which could then lead to more regressions,
as it won't be able to detect them in time. Changes to compute-hyperv are few
and strictly related to Hyper-V.