Install compute-hyperv in editable mode only on master
Because of the change in Nova that enforces drivers to be in the
nova.virt namespace and the way in which compute-hyperv is
currently installed, it has to be installed on editable mode
on master.

This will not work on stable branches, where compute-hyperv and
networking-hyperv are expected to be installed in the same place.